Addition of the logic to fetch all rentals + a specific rental

+++ src/main/java/com/openclassrooms/p3/payload/request/RentalUpdateRequest.java
@@ -1,5 +1,7 @@
 package com.openclassrooms.p3.payload.request;
 
+import java.math.BigDecimal;
+
 import org.springframework.lang.Nullable;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -9,9 +11,9 @@ import jakarta.validation.constraints.NotBlank;
  * Request payload containing the Form Data for updating a rental.
  */
 public record RentalUpdateRequest(
-                @NotBlank(message = "Name cannot be blank or null") String name,
-                @NotBlank(message = "Password cannot be blank or null") Integer surface,
-                @NotBlank(message = "Password cannot be blank or null") Double price,
-                @Nullable MultipartFile picture,
-                @NotBlank(message = "Password cannot be blank or null") String description) {
+        @NotBlank(message = "Name cannot be blank or null") String name,
+        @NotBlank(message = "Password cannot be blank or null") Integer surface,
+        @NotBlank(message = "Password cannot be blank or null") BigDecimal price,
+        @Nullable MultipartFile picture,
+        @NotBlank(message = "Password cannot be blank or null") String description) {
 }
